+++ test/integration/about_test.rb
@@ -1,4 +1,4 @@
-require 'test_helper'
+require 'integration_test_helper'
 
 class AboutIntegrationTest < ActionDispatch::IntegrationTest
   test "about page" do
