+++ test/integration/audit_test.rb
@@ -1,4 +1,4 @@
-require 'test_helper'
+require 'integration_test_helper'
 
 class AuditIntegrationTest < ActionDispatch::IntegrationTest
   test "index page" do
@@ -8,14 +8,5 @@ class AuditIntegrationTest < ActionDispatch::IntegrationTest
   test "audit content" do
     visit audits_path
     assert has_content?("updated Host"), "expected 'updated Host' but it doesn't exist"
-  end
-
-  test "show audit with diff" do
-    Capybara.current_driver = Capybara.javascript_driver
-    login_admin
-    audit = FactoryGirl.create(:audit, :with_diff)
-    visit audit_path(audit)
-    assert has_content?('-old')
-    assert has_content?('+new')
   end
 end
