+++ test/integration/auth_source_ldap_test.rb
@@ -1,4 +1,4 @@
-require 'test_helper'
+require 'integration_test_helper'
 
 class AuthSourceLdapIntegrationTest < ActionDispatch::IntegrationTest
   test "index page" do
