+++ test/integration/bookmark_test.rb
@@ -1,4 +1,4 @@
-require 'test_helper'
+require 'integration_test_helper'
 
 class BookmarkIntegrationTest < ActionDispatch::IntegrationTest
   test "index page" do
