+++ test/integration/common_parameter_test.rb
@@ -1,4 +1,4 @@
-require 'test_helper'
+require 'integration_test_helper'
 
 class CommonParameterIntegrationTest < ActionDispatch::IntegrationTest
   test "index page" do
