+++ test/integration/compute_profile_test.rb
@@ -1,4 +1,4 @@
-require 'test_helper'
+require 'integration_test_helper'
 
 class ComputeProfileIntegrationTest < ActionDispatch::IntegrationTest
   setup do
