+++ test/integration/config_report_test.rb
@@ -1,4 +1,4 @@
-require 'test_helper'
+require 'integration_test_helper'
 
 class ConfigReportIntegrationTest < ActionDispatch::IntegrationTest
   def setup
