+++ test/integration/hostgroup_test.rb
@@ -1,4 +1,4 @@
-require 'test_helper'
+require 'integration_test_helper'
 
 class HostgroupIntegrationTest < ActionDispatch::IntegrationTest
   test "index page" do
@@ -58,59 +58,5 @@ class HostgroupIntegrationTest < ActionDispatch::IntegrationTest
     assert page.has_link?('Parameters', :href => '#params')
     click_link 'Parameters'
     assert page.has_no_selector?('#params .has-error')
-  end
-
-  describe 'JS enabled tests' do
-    setup do
-      @old_driver = Capybara.current_driver
-      Capybara.current_driver = Capybara.javascript_driver
-      login_admin
-    end
-
-    teardown do
-      Capybara.current_driver = @old_driver
-    end
-
-    test 'submit updates taxonomy' do
-      group = FactoryGirl.create(:hostgroup, :with_puppetclass)
-      new_location = FactoryGirl.create(:location)
-
-      visit edit_hostgroup_path(group)
-      page.find(:css, "a[href='#locations']").click()
-      select_from_list 'hostgroup_location_ids', new_location
-
-      click_button "Submit"
-      #wait for submit to finish
-      page.find('#search-form')
-
-      group.locations.reload
-
-      assert_not_nil group.locations.first{ |l| l.name == new_location.name }
-    end
-
-    test 'parameters change after parent update' do
-      group = FactoryGirl.create(:hostgroup)
-      group.group_parameters << GroupParameter.create(:name => "x", :value => "original")
-      child = FactoryGirl.create(:hostgroup)
-
-      visit clone_hostgroup_path(child)
-      assert page.has_link?('Parameters', :href => '#params')
-      click_link 'Parameters'
-      assert page.has_no_selector?("#inherited_parameters #name_x")
-
-      click_link 'Hostgroup'
-      select2(group.name, :from => 'hostgroup_parent_id')
-      wait_for_ajax
-
-      click_link 'Parameters'
-      assert page.has_selector?("#inherited_parameters #name_x")
-    end
-  end
-
-  private
-
-  def select_from_list(list_id, item)
-    span = page.all(:css, "#ms-#{list_id} .ms-selectable ul li span").first{ |i| i.text == item.name }
-    span.click
   end
 end
