+++ test/integration/media_test.rb
@@ -1,4 +1,4 @@
-require 'test_helper'
+require 'integration_test_helper'
 
 class MediaIntegrationTest < ActionDispatch::IntegrationTest
   test "index page" do
