+++ test/integration/operatingsystem_test.rb
@@ -1,4 +1,4 @@
-require 'test_helper'
+require 'integration_test_helper'
 
 class OperatingsystemIntegrationTest < ActionDispatch::IntegrationTest
   test "index page" do
