+++ test/integration/organization_test.rb
@@ -1,4 +1,4 @@
-require 'test_helper'
+require 'integration_test_helper'
 
 class OrganizationIntegrationTest < ActionDispatch::IntegrationTest
   def setup
@@ -26,7 +26,7 @@ class OrganizationIntegrationTest < ActionDispatch::IntegrationTest
   # context - creating when all hosts are assigned
   test "create new page when all hosts are assigned a organization" do
     Host.update_all(:organization_id => Organization.first.id)
-    assert !has_selector?("div.alert", :text => "with no organization assigned")
+    assert has_no_selector?("div.alert", :text => "with no organization assigned")
     assert_new_button(organizations_path,"New Organization",new_organization_path)
     fill_in "organization_name", :with => "Finance"
     assert_submit_button(organizations_path)
@@ -65,24 +65,6 @@ class OrganizationIntegrationTest < ActionDispatch::IntegrationTest
     click_link "Proceed to Edit"
     assert_current_path edit_organization_path(Organization.unscoped.order(:id).last)
     assert page.has_selector?('h1', :text => "Edit"), "Edit was expected in the <h1> tag, but was not found"
-  end
-
-  test "multiselect does not add items that are filtered out" do
-    Capybara.current_driver = Capybara.javascript_driver
-    login_admin
-
-    visit edit_organization_path(taxonomies(:organization1))
-
-    wait_for_ajax
-    within "#content" do
-      click_link "Locations"
-      within "#locations" do
-        find(".ms-selection").assert_no_selector("li[selected='selected']")
-        find(".ms-selectable").find("input").set("Location 1")
-        find("a[data-original-title='Select All']").click
-        find(".ms-selection").find(".ms-selected").find("span").has_text? "Location 1"
-      end
-    end
   end
 
   # PENDING
