+++ test/integration/puppetclass_test.rb
@@ -1,4 +1,4 @@
-require 'test_helper'
+require 'integration_test_helper'
 
 class PuppetclassIntegrationTest < ActionDispatch::IntegrationTest
   test "edit page" do
