+++ test/integration/smart_proxy_test.rb
@@ -1,4 +1,4 @@
-require 'test_helper'
+require 'integration_test_helper'
 
 class SmartProxyIntegrationTest < ActionDispatch::IntegrationTest
   test "index page" do
