+++ test/integration/usergroup_test.rb
@@ -1,4 +1,4 @@
-require 'test_helper'
+require 'integration_test_helper'
 
 class UsergroupIntegrationTest < ActionDispatch::IntegrationTest
   def setup
