+++ test/integration/variable_lookup_key_test.rb
@@ -1,4 +1,4 @@
-require 'test_helper'
+require 'integration_test_helper'
 
 class VariableLookupKeyIntegrationTest < ActionDispatch::IntegrationTest
   test "index page" do
