+++ test/lib/middleware/catch_json_parse_errors_test.rb
@@ -1,4 +1,4 @@
-require 'test_helper'
+require 'integration_test_helper'
 require 'rest-client'
 require 'capybara/rails'
