+++ test/unit/fact_parser_test.rb
@@ -1,4 +1,5 @@
 require "test_helper"
+require 'minitest/mock'
 
 class FactParserTest < ActiveSupport::TestCase
   attr_reader :parser, :host
