+++ test/unit/filter_test.rb
@@ -1,4 +1,5 @@
 require 'test_helper'
+require 'minitest/mock'
 
 class FilterTest < ActiveSupport::TestCase
   test "#unlimited?" do
