+++ test/unit/host_status/build_status_test.rb
@@ -1,4 +1,5 @@
 require 'test_helper'
+require 'minitest/mock'
 
 class BuildStatusTest < ActiveSupport::TestCase
   def setup
