+++ test/unit/user_test.rb
@@ -1,5 +1,6 @@
 # encoding: UTF-8
 require 'test_helper'
+require 'minitest/mock'
 
 class UserTest < ActiveSupport::TestCase
   def setup
